+++ pkg/plugin/check_health.go
@@ -12,25 +12,26 @@ import (
 
 func checkDB(pathPrefix string, path string, options string) error {
 	if pathPrefix == "file:" || pathPrefix == "" {
-		_, err := os.Stat(path)
+		fileInfo, err := os.Stat(path)
 		if errors.Is(err, os.ErrNotExist) {
 			return fmt.Errorf("no file exists at the file path")
 		} else if err != nil {
-			return err
+			return fmt.Errorf("error checking path: %v", err)
+		} else if fileInfo.IsDir() {
+			return fmt.Errorf("the provided path is a directory instead of a file")
 		}
 	}
 
 	db, err := sql.Open("sqlite", pathPrefix+path+"?"+options)
 	if err != nil {
-		return err
+		return fmt.Errorf("error opening %s%s: %v", pathPrefix, path, err)
 	}
 	defer db.Close()
 
-	sth, err := db.Exec("pragma schema_version;")
+	_, err = db.Exec("pragma schema_version;")
 	if err != nil {
-		return err
+		return fmt.Errorf("error checking for valid SQLite file: %v", err)
 	}
-	fmt.Println(sth)
 
 	return nil
 }